backend/requirements.txt
# FastAPI ve web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Veri işleme
pandas==2.1.3
numpy==1.24.3
langdetect==1.0.9

# ML ve Embedding
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2

# Vektör veritabanı
pinecone-client==2.2.4

# LLM entegrasyonu
langchain==0.0.350
langchain-community==0.0.1
groq==0.4.2

# HTTP istekleri
requests==2.31.0
aiohttp==3.9.1

# Veritabanı
sqlite3
sqlalchemy==2.0.23

# Çeviri
googletrans==4.0.0rc1

# Yardımcı araçlar
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Test
pytest==7.4.3
pytest-asyncio==0.21.1 